Game System work correctly

+++ lib/business/services/GameService.dart
@@ -1,4 +1,5 @@
 import 'package:takwira/business/services/TeamService.dart';
+import 'package:takwira/presentation/managers/ChatManager.dart';
 
 import '../../domain/entities/Game.dart';
 import '../../domain/repositories/IGameRepository.dart';
@@ -44,11 +45,12 @@ class GameService implements IGameService {
     try {
       // Retrieve game details
       Game game = await getGameDetails(gameId);
-
-      // Cancel the current game for the home team
-      await TeamService().cancelCurrentGameFromTeam(game.homeTeam);
-
-      await _gameRepository.deleteGame(gameId);
+      if (game.chatId != null) {
+        // Cancel the current game for the home team
+        await TeamService().cancelCurrentGameFromTeam(game.homeTeam);
+        await ChatManager().deleteChat(game.chatId!);
+        await _gameRepository.deleteGame(gameId);
+      }
     } catch (e) {
       throw Exception('Failed to delete game: $e');
     }
@@ -66,4 +68,42 @@ class GameService implements IGameService {
       throw Exception('Failed to complete game: $e');
     }
   }
+
+  @override
+  Future<bool> updateGameDate(String gameId, DateTime newDate) async {
+    try {
+      Game game = await getGameDetails(gameId);
+      game.gameDate = newDate;
+      await updateGame(game);
+      return true;
+    } on Exception {
+      return false;
+    }
+  }
+
+  @override
+  Future<bool> updateGameStadium(String gameId, String stadiumId) async {
+      try {
+        Game game = await getGameDetails(gameId);
+        game.stadiumId = stadiumId;
+        await updateGame(game);
+        return true;
+      } on Exception {
+        return false;
+      }
+    }
+
+
+  @override
+  Future<bool> cancelGameStadium(String gameId) async {
+    try {
+      Game game = await getGameDetails(gameId);
+      game.stadiumId = null;
+      await updateGame(game);
+      return true;
+    } on Exception {
+      return false;
+    }
+  }
+
 }
